Move addQueueTimes to QueueModel
