+++ packages/server/src/queue/queue.entity.ts
@@ -10,11 +10,18 @@ import {
   ManyToOne,
   OneToMany,
   PrimaryGeneratedColumn,
+  MoreThanOrEqual,
+  LessThanOrEqual,
 } from 'typeorm';
 import { CourseModel } from '../course/course.entity';
 import { UserModel } from '../profile/user.entity';
 import { QuestionModel } from '../question/question.entity';
 import { OfficeHourModel } from '../course/office-hour.entity';
+
+interface TimeInterval {
+  startTime: Date;
+  endTime: Date;
+}
 
 @Entity('queue_model')
 export class QueueModel extends BaseEntity {
@@ -44,8 +51,6 @@ export class QueueModel extends BaseEntity {
   @ManyToMany((type) => UserModel, (user) => user.queues)
   @JoinTable()
   staffList: UserModel[];
-
-  // If you need to add time to queues check out this commit: 995e82991587b2077d342b1df87a2665a21c3492
 
   @Exclude()
   @OneToMany((type) => OfficeHourModel, (oh) => oh.queue)
@@ -79,6 +84,70 @@ export class QueueModel extends BaseEntity {
 
   startTime: Date;
   endTime: Date;
+
+  public async addQueueTimes() {
+    const now = new Date();
+
+    const officeHours = await this.getOfficeHours(this.id);
+    const timeIntervals = await this.generateMergedTimeIntervals(officeHours);
+    const currTime = timeIntervals.find(
+      (group) => group.startTime <= now && group.endTime >= now,
+    );
+
+    this.startTime = currTime.startTime;
+    this.endTime = currTime.endTime;
+  }
+
+  private async getOfficeHours(queueId: number): Promise<OfficeHourModel[]> {
+    const now = new Date();
+
+    const lowerBound = new Date(now);
+    lowerBound.setUTCHours(now.getUTCHours() - 24);
+    lowerBound.setUTCHours(0, 0, 0, 0);
+
+    const upperBound = new Date(now);
+    upperBound.setUTCHours(now.getUTCHours() + 24);
+    upperBound.setUTCHours(0, 0, 0, 0);
+
+    return await OfficeHourModel.find({
+      where: [
+        {
+          queueId: queueId,
+          startTime: MoreThanOrEqual(lowerBound),
+          endTime: LessThanOrEqual(upperBound),
+        },
+      ],
+      order: {
+        startTime: 'ASC',
+      },
+    });
+  }
+
+  private generateMergedTimeIntervals(
+    officeHours: OfficeHourModel[],
+  ): TimeInterval[] {
+    const timeIntervals: TimeInterval[] = [];
+    officeHours.forEach((officeHour) => {
+      if (
+        timeIntervals.length == 0 ||
+        officeHour.startTime > timeIntervals[timeIntervals.length - 1].endTime
+      ) {
+        timeIntervals.push({
+          startTime: officeHour.startTime,
+          endTime: officeHour.endTime,
+        });
+        return;
+      }
+
+      const prevGroup = timeIntervals[timeIntervals.length - 1];
+      timeIntervals[timeIntervals.length - 1].endTime =
+        officeHour.endTime > prevGroup.endTime
+          ? officeHour.endTime
+          : prevGroup.endTime;
+    });
+
+    return timeIntervals;
+  }
 
   // TODO: eventually figure out how staff get sent to FE as well
 }
